fix(r188/mm): failed-dealloc retention, framebuffer containment, page-cache writeback oracle, TLB normalization
R188 standalone audit remediation — allocator / MM leg.

U26-2: never silently discard a failed deallocation. U26-3: normalize TLB
shootdown PCIDs at the callee boundary. U24-2: page-cache resident write paths
make the writeback oracle non-vacuous. memory.rs gains the framebuffer
region alignment/containment regression; the admitted/heap_admission/oom/page_table
paths gain the associated fallible-admission and diagnostics hardening.

Covered by the mm hosted suite (25/25).

+++ kernel/mm/oom_killer.rs
@@ -42,8 +42,8 @@ type SnapshotFn = fn() -> Option<OomProcessInfo>;
 type KillFn = fn(ProcessId, u64, i32) -> bool;
 type TimestampFn = fn() -> u64;
 /// R106-8: Callback for emitting tamper-evident audit events.
-/// Args: (pid: u32, uid: u32, nr_pages_needed: u64, rss_pages: u64, oom_score_adj: i64, timestamp: u64)
-type AuditEmitFn = fn(u32, u32, u64, u64, i64, u64);
+/// Args: (pid: u64, uid: u32, nr_pages_needed: u64, rss_pages: u64, oom_score_adj: i64, timestamp: u64)
+type AuditEmitFn = fn(u64, u32, u64, u64, i64, u64);
 
 struct Callbacks {
     snapshot: Option<SnapshotFn>,
@@ -291,7 +291,7 @@ fn emit_audit(
     // and verified for integrity via sys_audit_export.
     if let Some(emit) = audit_cb {
         emit(
-            victim.pid as u32,
+            victim.pid as u64,
             victim.uid,
             needed as u64,
             victim.rss_pages,
